smoketest: support both Automatic and Timed Login smoketest
We should check both login paths, in case one of them got broken.
This also switches default login type to Automatic.

Default Timed login timeout is 30 seconds, which is too much - we'll boot
with 10 seconds delay - this should be enough to verify that timedlogin
works and make screenshots of login indicator